reorganize group components
