+++ frontend/src/components/Groups/GroupPage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const GroupPage = () => {
+  return (
+    <div>GroupPage</div>
+  )
+}
+
+export default GroupPage
